Remove COM_*_rel, R_*_rel functions; no longer used Replace R_move_abs -> R_pos_abs Replace R_move_abs + R_cont_abs -> R_line_abs Add D_pos_abs, D_pos_rel Eliminate dual-use of D_{move,cont}_{abs,rel} Update d.graph to use D_pos_abs for move and D_line_abs for move/draw Merge D_symbol and D_symbol2 Replace R_* -> D_* in D_symbol Update d.vect to handle D_symbol changes Eliminate integer coordinates in soft_text functions
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@33270 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/driver/driverlib.h
@@ -14,14 +14,14 @@ void font_list(char ***, int *, int);
 
 /* Text2.c */
 void drawchar(double, double, double, double, unsigned char);
-void soft_text_ext(int, int, double, double, double, const char *);
+void soft_text_ext(double, double, double, double, double, const char *);
 void get_text_ext(double *, double *, double *, double *);
-void soft_text(int, int, double, double, double, const char *);
-void onechar(int, int, double, double, double, unsigned char);
+void soft_text(double, double, double, double, double, const char *);
+void onechar(double, double, double, double, double, unsigned char);
 
 /* Text3.c */
-void soft_text_freetype(int, int, double, double, double, const char *);
-void soft_text_ext_freetype(int, int, double, double, double, const char *);
+void soft_text_freetype(double, double, double, double, double, const char *);
+void soft_text_ext_freetype(double, double, double, double, double, const char *);
 void get_text_ext_freetype(double *, double *, double *, double *);
 
 /* font2.c */
